Update and deprecate pgsql factory due to core changes
This factory has not resulted in simplicity, and also the use of multiple DataSource instances governed by a single place. Instead, an example similar to mssql-hikari should be developed to replace instances of this factory and factory builder. Apps know what db they connect to and don't need a connection abstraction layer, just an abstraction for a connection itself (or connected DataSource)